requirements.txt
langchain-anthropic
langchain
langchain_community
python-dotenv
streamlit
ipykernel
wikipedia
bs4
langchain-text-splitters
voyageai
langchain_voyageai
chromadb
faiss-cpu
langsmith
langchain-groq
fastapi
langserve
sse_starlette
langchain_aws
validators
youtube_transcript_api
unstructured
wikipedia
pymupdf
pdfplumber
arxiv
duckduckgo-search
jq